adds documents upload to animal extra images moves file util function from components > utils to root utils
