+++ frontend/src/animals/AnimalForm.js
@@ -4,7 +4,7 @@ import { navigate, useNavigationPrompt, useQueryParams } from 'raviger';
 import { Form, Formik } from "formik";
 import { ButtonGroup, Card, Col, Image, Form as BootstrapForm } from "react-bootstrap";
 import * as Yup from 'yup';
-import { AddressSearch, DateTimePicker, DropDown, ImageUploader, TextInput } from '../components/Form.js';
+import { AddressSearch, DateTimePicker, DropDown, FileUploader, ImageUploader, TextInput } from '../components/Form.js';
 import { catAgeChoices, dogAgeChoices, horseAgeChoices, otherAgeChoices, catColorChoices, dogColorChoices, horseColorChoices, otherColorChoices, sexChoices, dogSizeChoices, catSizeChoices, horseSizeChoices, otherSizeChoices, statusChoices, reportedStatusChoices, unknownChoices } from './constants';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faArrowAltCircleLeft, faMinusSquare } from '@fortawesome/free-solid-svg-icons';
@@ -860,7 +860,7 @@ const AnimalForm = (props) => {
                       </span>
                     :""}
                     <div className="mb-2" hidden={!id}>
-                      <ImageUploader
+                      <FileUploader
                         value={extra_images}
                         id="extra_images"
                         name="extra_images"
