+++ frontend/src/utils/files.js
@@ -1,0 +1,9 @@
+import { IMAGE_TYPES } from '../constants';
+
+export function isImageFile(filename) {
+  return IMAGE_TYPES.includes(filename.split('.').pop());
+}
+
+export function getFileNameFromUrl(url) {
+  return url.split('/').pop().split('.')[0];
+}
